fix docs-theme/docs-app
