+++ packages/docs-app/src/examples/datetime-examples/index.ts
@@ -18,5 +18,4 @@ export * from "./dateInputExample";
 export * from "./datePickerExample";
 export * from "./dateRangeInputExample";
 export * from "./dateRangePickerExample";
-export * from "./dateTimePickerExample";
 export * from "./timePickerExample";
